Use ReTest.jl for tests
