+++ .gitignore
@@ -7,6 +7,7 @@ _*.jl
 *.jl.mem
 
 ./logging_config.toml
+test/*.log
 
 ## Autogenerated code during the documentation process
 generated*.md
@@ -39,6 +40,8 @@ docs/site/
 Manifest.toml
 .vscode
 *.h5
+# ctags
+tags
 
 ################################################################################
 #                              Operating systems                               #
