+++ test/runtests.jl
@@ -1,121 +1,4 @@
-using Test
-using Logging
-import Dates
-import TerminalLoggers: TerminalLogger
-import TimeSeries
-import UUIDs
-import JSON3
-import HDF5
-using DataStructures: SortedDict
-using DataFrames
-using Random
-using ProgressLogging
+using InfrastructureSystems
 
-import InfrastructureSystems
-
-import Aqua
-Aqua.test_unbound_args(InfrastructureSystems)
-Aqua.test_undefined_exports(InfrastructureSystems)
-Aqua.test_ambiguities(InfrastructureSystems)
-Aqua.test_stale_deps(InfrastructureSystems)
-Aqua.test_deps_compat(InfrastructureSystems)
-
-const IS = InfrastructureSystems
-const BASE_DIR =
-    abspath(joinpath(dirname(Base.find_package("InfrastructureSystems")), ".."))
-const DATA_DIR = joinpath(BASE_DIR, "test", "data")
-const FORECASTS_DIR = joinpath(DATA_DIR, "time_series")
-
-const LOG_FILE = "infrastructure-systems.log"
-
-include("common.jl")
-
-"""
-Copied @includetests from https://github.com/ssfrr/TestSetExtensions.jl.
-Ideally, we could import and use TestSetExtensions.  Its functionality was broken by changes
-in Julia v0.7.  Refer to https://github.com/ssfrr/TestSetExtensions.jl/pull/7.
-"""
-
-"""
-Includes the given test files, given as a list without their ".jl" extensions.
-If none are given it will scan the directory of the calling file and include all
-the julia files.
-"""
-macro includetests(testarg...)
-    if length(testarg) == 0
-        tests = []
-    elseif length(testarg) == 1
-        tests = testarg[1]
-    else
-        error("@includetests takes zero or one argument")
-    end
-
-    quote
-        tests = $tests
-        rootfile = @__FILE__
-        if length(tests) == 0
-            tests = readdir(dirname(rootfile))
-            tests = filter(
-                f ->
-                    startswith(f, "test_") && endswith(f, ".jl") && f != basename(rootfile),
-                tests,
-            )
-        else
-            tests = map(f -> string(f, ".jl"), tests)
-        end
-        println()
-        for test in tests
-            print(splitext(test)[1], ": ")
-            include(test)
-            println()
-        end
-    end
-end
-
-function get_logging_level_from_env(env_name::String, default)
-    level = get(ENV, env_name, default)
-    return IS.get_logging_level(level)
-end
-
-function run_tests()
-    logging_config_filename = get(ENV, "SIIP_LOGGING_CONFIG", nothing)
-    if logging_config_filename !== nothing
-        config = IS.LoggingConfiguration(logging_config_filename)
-    else
-        config = IS.LoggingConfiguration(;
-            filename = LOG_FILE,
-            file_level = Logging.Info,
-            console_level = Logging.Error,
-        )
-    end
-    console_logger = TerminalLogger(config.console_stream, config.console_level)
-
-    IS.open_file_logger(config.filename, config.file_level) do file_logger
-        levels = (Logging.Info, Logging.Warn, Logging.Error)
-        multi_logger =
-            IS.MultiLogger([console_logger, file_logger], IS.LogEventTracker(levels))
-        global_logger(multi_logger)
-
-        if !isempty(config.group_levels)
-            IS.set_group_levels!(multi_logger, config.group_levels)
-        end
-
-        @time @testset "Begin Systems tests" begin
-            @includetests ARGS
-        end
-
-        @test length(IS.get_log_events(multi_logger.tracker, Logging.Error)) == 0
-
-        @info IS.report_log_summary(multi_logger)
-    end
-end
-
-logger = global_logger()
-
-try
-    run_tests()
-finally
-    # Guarantee that the global logger is reset.
-    global_logger(logger)
-    nothing
-end
+include("InfrastructureSystemsTests.jl")
+run_tests()
